big update
printing page (add trxn, co maker and reason)
loan new inputs (reason, co maker)
adjust loanlist viewing

+++ ajax/ajaxowner.php
@@ -3,6 +3,12 @@
 	if(!isset($_GET['module'])){
 		echo '<script type = "text/javascript">window.location.replace("/loan/");</script>';
 	}
+	session_start();
+	$access = "SELECT * FROM user where account_id = '$_SESSION[acc_id]'";
+	$access = $conn->query($access)->fetch_object();
+	if(!isset($_GET['module'])){
+		echo '<script type = "text/javascript">window.location.replace("/loan/");</script>';
+	}
 ?>
 <?php
 	if(isset($_GET['loanlist'])){
@@ -11,12 +17,12 @@
 		$search = mysqli_real_escape_string($conn, $_GET['loanlist']);
 		$search = str_replace(",", "", $search);
 		$where = " and (fname like '%$search%' or mname like '%$search%' or lname like '%$search%' or principal like '%$search%' or duration like '%$search%' or type like '%$search%') ";
-		if($search == ""){
+		if(empty($_GET['loanlist'])){
 			$where = "";
 		}
 		$counter = "SELECT count(*) as total FROM customer as a,loan as b,breakdown as c where a.customer_id = b.customer_id and b.loan_id = c.loan_id and c.state = '0' " . $where . " group by b.loan_id";
 		$counter2 = $conn->query($counter)->fetch_assoc();
-		$perpage = 2;
+		$perpage = 15;
 		$totalPages = ceil($counter2['total'] / $perpage);
 		if(!isset($_GET['page'])){
 		    $_GET['page'] = 0;
@@ -31,22 +37,41 @@
 		$startArticle = ($_GET['page'] - 1) * $perpage;
 		$list = "SELECT * FROM customer as a,loan as b,breakdown as c where a.customer_id = b.customer_id and b.loan_id = c.loan_id and c.state = '0' " . $where . " group by b.loan_id LIMIT " . $startArticle . ', ' . $perpage;
 		$res = $conn->query($list);
+		$totalloan = 0; $totalinterest = 0; $total = 0; $moninte = 0;
 		if($res->num_rows > 0){
 			$num = 0;
 			while ($row = $res->fetch_assoc()) {
 				$gerate = "SELECT ".strtolower($row['type']) ." as rate FROM rate";
 				$gerate = $conn->query($gerate)->fetch_assoc();
 				$num += 1;	
+				$totalloan += $row['principal'];
+				$totalinterest += ($row['principal'] * $row['rate']) * $row['duration'];
+				$moninte += ($row['principal'] * $row['rate']);
+				$total += ( (($row['principal'] * $row['rate']) * $row['duration']) +  $row['principal'] );
 				echo '<tr>';
 				echo '<td>' . $num . '</td>';
 				echo '<td>' . $row['fname'] . ' ' . $row['mname'] . ' ' . $row['lname'] . ' ( ' . $row['customer_id'] . ' )</td>';
 				echo '<td>₱ ' . number_format($row['principal'],2) . '</td>';
-				echo '<td>₱ ' . number_format($row['principal'] * $row['rate'],2) . '</td>';
-				echo '<td>₱ ' . number_format(str_replace(",", "", number_format($row['principal'] * $row['rate'],2)) + str_replace(",", "", number_format($row['principal'],2)),2) . '</td>';
+				echo '<td>₱ ' . number_format( ($row['principal'] * $row['rate']), 2) . '</td>';
+				echo '<td>₱ ' . number_format( ($row['principal'] * $row['rate']) * $row['duration'], 2) . '</td>';
+				echo '<td>₱ ' . number_format(str_replace(",", "", number_format(($row['principal'] * $row['rate']) * $row['duration'],2)) + str_replace(",", "", number_format($row['principal'],2)),2) . '</td>';
+				echo '<td>' . date("M j, Y", strtotime($row['startdate'])) . '</td>';
 				echo '<td>' . $row['duration'] . ' - ' . $row['type'] . '</td>';
-				echo '<td><a href = "?module=loan&action=view&id='.$row['loan_id'].'" class = "btn btn-sm btn-primary"> View Details </a></td>';
+				echo 
+					'<td>
+						<a href = "loan/view/'.$row['loan_id'].'" class = "btn btn-sm btn-primary" data-toggle="tooltip" title="View"><span class = "icon-search"></span></a>
+						<a onclick = "payment('.$row['loan_id'].');" class = "btn btn-sm btn-success" data-toggle="tooltip" title="Payment"><span>₱</span></a>';
+				echo '</td>';
 				echo '</tr>';
 			}
+			echo '<tr>';
+				echo '<td></td><td align = "right"> <b>Total:</b></td>';
+				echo '<td>₱ ' . number_format($totalloan,2) . '</td>';
+				echo '<td>₱ ' . number_format($moninte,2) . '</td>';
+				echo '<td>₱ ' . number_format($totalinterest,2) . '</td>';
+				echo '<td>₱ ' . number_format($total,2) . '</td>';
+				echo '<td></td><td></td><td></td>';
+			echo '</tr>';
 		}else{
 			echo '<tr><td colspan = "6" align = "center"> <h5> No Record Found </h5></td></tr>';
 		}
@@ -73,26 +98,46 @@
 	        		$payment = $conn->query($payment)->fetch_object();
 	        		echo '<div class = "form-group"><label>Principal Balance: ₱ ' . number_format($payment->principal - $payment->sumpayprincipal,2) . ' </label></div>';
 	        		echo '<hr>';
-	        		echo '<div class = "form-group"><label>Interest Balance (* '. $payment->rate .' ): ₱ ' . number_format(($payment->principal * $payment->rate) - $payment->sumpayinterest,2) . ' </label></div>';
+	        		echo '<div class = "form-group"><label>Interest Balance (* '. $payment->rate .' ): ₱ ' . number_format((($payment->principal * $payment->rate) * $payment->duration) - $payment->sumpayinterest,2) . ' </label></div>';
 	        		echo '<hr>';
 	        		$diff=date_diff(date_create($payment->due),date_create(date("Y-m-d")));
 					if($diff->format("%R%") == '+' && $diff->format("%a%") > 0 && $payment->due <= date('Y-m-d') && $payment->state == 0){
 						$penalty = number_format((($payment->principal + ($payment->principal * $payment->rate)) * $gerate['penalty']) * $diff->format("%a%"),2);
-						echo '<div class = "form-group"><label>Penalty Balance (* '. $gerate['penalty'] .' ): ₱ '. $penalty . ' <font color = "red">( '. $diff->format("%a%") . ' day/s) </font> </label></div>';
-	        			echo '<hr>';
+						if( (str_replace(",", "", $penalty) - $payment->sumpaypenalty) > 0 ){	
+							echo '<div class = "form-group"><label>Penalty Balance (* '. $gerate['penalty'] .' ): ₱ '. $penalty . ' <font color = "red">( '. $diff->format("%a%") . ' day/s) </font> </label></div>';
+		        			echo '<hr>';
+						}
 					}
 	        	?>
 	        	<form role="form" action = "" method = "post">
 	        		<div class="form-group">
-	            		<label>Principal </label>
+	        			<label>Type of Payment <font color = "red"> * </font></label>
+	        			<select class="form-control input-sm" required name = "paytype" onchange="paytypex(this.value)">
+	        				<option> - - - - - </option>
+	        				<option value = "Cash">Cash</option>
+	        				<option value = "Check">Check</option>
+	        			</select>
+	        		</div>
+	        		<div id = "check" style="display: none;">
+	        			<div class="form-group">
+	        				<label>Check No. <font color = "red"> * </font></label>
+	        				<input type = "text" name = "checknum" class="form-control input-sm" placeholder = "Enter check no.">
+	        			</div>
+	        			<div class="form-group">
+	        				<label>Check Date <font color = "red"> * </font></label>
+	        				<input type = "date" name = "check" class="form-control input-sm">
+	        			</div>
+	        		</div>
+	        		<div class="form-group">
+	            		<label>Principal <font color = "red"> * </font></label>
 	            		<input value = '0' type = "text" name = "prin" class="form-control input-sm" placeholder = "Enter amount" pattern = "[.0-9]*" required>
 		            </div>
 		            <div class="form-group">
-	            		<label>Interest </label>
+	            		<label>Interest <font color = "red"> * </font></label>
 	            		<input value = '0' type = "text" name = "inte" class="form-control input-sm" placeholder = "Enter amount" pattern = "[.0-9]*" required>
 	            	</div>
 	            	<div class="form-group">
-	            		<label>Penalty </label>
+	            		<label>Penalty <font color = "red"> * </font></label>
 	            		<input value = '0' type = "text" name = "penal" class="form-control input-sm" placeholder = "Enter amount" pattern = "[.0-9]*">
 	            	</div>
 	            	<button type="submit" name = "paysub" class="btn btn-success btn-block" onclick="return confirm('Are you sure?');"><span class ="icon-checkmark"></span> Add payment</button>
@@ -153,7 +198,7 @@ if(isset($_GET['amount']) && isset($_GET['type']) && isset($_GET['duration']) &&
 	</div>
 	<div class="col-md-2 col-xs-6" style="border-right: 1px solid #eee;">
 		<label><u>Interest</u></label>
-		<i><p style="margin-left: 20px;">₱ <?php echo number_format($amount * $gerate['rate'],2);?></p></i>
+		<i><p style="margin-left: 20px;">₱ <?php echo number_format($amount * $gerate['rate'] * $duration,2);?></p></i>
 	</div>
 	<div class="col-md-2 col-xs-6" style="border-right: 1px solid #eee;">
 		<label><u>Duration - Type </u></label>
@@ -161,7 +206,7 @@ if(isset($_GET['amount']) && isset($_GET['type']) && isset($_GET['duration']) &&
 	</div>
 	<div class="col-md-2 col-xs-6" style="border-right: 1px solid #eee;">
 		<label><u>Total </u></label>
-		<i><p style="margin-left: 20px;">₱ <?php echo number_format(($amount * $gerate['rate']) + $amount,2);?></p></i>
+		<i><p style="margin-left: 20px;">₱ <?php echo number_format((($amount * $gerate['rate']) * $duration) + $amount,2);?></p></i>
 	</div>
 	<div class="col-md-2 col-xs-6" style="border-right: 1px solid #eee;">
 		<label><u>Due Date </u></label>
@@ -187,51 +232,39 @@ if(isset($_GET['amount']) && isset($_GET['type']) && isset($_GET['duration']) &&
 	<div class="col-md-2 col-xs-3">
 		<label><u>Principal Amount</u></label>
 		<?php
-			$brkamnt = number_format($amount/$duration,2);
-			$brkamnt = str_replace(",", "", $brkamnt);
-			$total = 0;
 			$totalprin = 0;
 			for($i = 0; $i < $duration; $i++){		
-				$total += $brkamnt;
-				if($i == ($duration - 1)){
-					$brkamnt += $amount - $total;
+				if($i > 0) {
+					echo '<i><p style="margin-left: 20px;"> - </p></i>';
+					continue;	
 				}
-				$totalprin += $brkamnt;
-				echo '<i><p style="margin-left: 20px;">₱ '.number_format($brkamnt,2).'</p></i>';
+				$totalprin = $amount;
+				echo '<i><p style="margin-left: 20px;">₱ '.number_format($amount,2).'</p></i>';
 			}
 		?>	
 	</div>
 	<div class="col-md-2 col-xs-3">
 		<label><u>Interest Amount</u></label>
 		<?php
-			$brkamnt = number_format((($amount * $gerate['rate']))/$duration,2);
-			$brkamnt = str_replace(",", "", $brkamnt);
-			$total = 0;
 			$totalinte = 0;
 			for($i = 0; $i < $duration; $i++){		
-				$total += $brkamnt;
-				if($i == ($duration - 1)){
-					$brkamnt += (($amount * $gerate['rate'])) - $total;
-				}
-				$totalinte += $brkamnt;
-				echo '<i><p style="margin-left: 20px;">₱ '.number_format($brkamnt,2).'</p></i>';
+				$totalinte += ($amount * $gerate['rate']);
+				echo '<i><p style="margin-left: 20px;">₱ '.number_format(($amount * $gerate['rate']),2).'</p></i>';
 			}
 		?>		
 	</div>
 	<div class="col-md-2 col-xs-3">
 		<label><u>Total Amount</u></label>
 		<?php
-			$brkamnt = number_format((($amount * $gerate['rate']) + $amount)/$duration,2);
-			$brkamnt = str_replace(",", "", $brkamnt);
-			$total = 0;
 			$totalamnt = 0;
 			for($i = 0; $i < $duration; $i++){		
-				$total += $brkamnt;
-				if($i == ($duration - 1)){
-					$brkamnt += (($amount * $gerate['rate']) + $amount) - $total;
+				if($i < 1){
+					$amntx = (($amount * $gerate['rate']) + $amount);
+				}else{
+					$amntx = ($amount * $gerate['rate']);
 				}
-				$totalamnt += $brkamnt;
-				echo '<i><p style="margin-left: 20px;"><b>₱ '.number_format($brkamnt,2).'</b></p></i>';
+				$totalamnt += $amntx;
+				echo '<i><p style="margin-left: 20px;"><b>₱ '.number_format($amntx,2).'</b></p></i>';
 			}
 		?>	
 	</div>
